added standardized navigation logic to taskserverbase & made all tasks use it
