+++ all_seaing_autonomy/all_seaing_autonomy/roboboat/speed_challenge.py
@@ -51,18 +51,10 @@ class SpeedChallenge(TaskServerBase):
         )
 
 
-        self.follow_path_client = ActionClient(self, FollowPath, "follow_path")
         self.waypoint_marker_pub = self.create_publisher(
             MarkerArray, "waypoint_markers", 10
         )
 
-        self.declare_parameter("xy_threshold", 1.0)
-        self.declare_parameter("theta_threshold", 180.0)
-        self.declare_parameter("goal_tol", 0.5)
-        self.declare_parameter("obstacle_tol", 50)
-        self.declare_parameter("choose_every", 10)
-        self.declare_parameter("use_waypoint_client", False)
-        self.declare_parameter("planner", "astar")
         self.declare_parameter("probe_distance", 10)
         self.declare_parameter("adaptive_distance", 0.7)
         self.adaptive_distance = self.get_parameter("adaptive_distance").get_parameter_value().double_value
@@ -89,11 +81,6 @@ class SpeedChallenge(TaskServerBase):
         self.buoy_direction = (0,0)
         self.buoy_found = False
         self.following_guide = False
-        self.moved_to_point = False
-        self.waypoint_rejected = False
-        self.waypoint_aborted = False
-        self.active_future_request = 0 # active future requests
-        self.active_waypoint_request = 0 # keeps track of the number of active waypoint requests
         self.left_first = True # goes left of buoy first
 
         self.obstacles = None
@@ -381,101 +368,6 @@ class SpeedChallenge(TaskServerBase):
             goal_update_func=partial(self.update_point, "gate_wpt", partial(self.update_gate_wpt_pos, self.forward_dist_back)))
         # self.move_to_point(self.home_pos, busy_wait=True)
         return Task.Result(success=True)
-
-    def _send_goal(self, goal_msg):
-        self.follow_path_client.wait_for_server()
-        self.active_future_request += 1
-        self.send_goal_future = self.follow_path_client.send_goal_async(
-            goal_msg
-        )
-        self.send_goal_future.add_done_callback(self.follow_path_response_cb)
-
-    def move_to_point(self, point, is_stationary=False, busy_wait=False, exit_func=None, goal_update_func=None):
-        '''
-        Moves the boat to the specified position using the follow path action server.
-        If busy_wait=true, then the returned truth value indicates success of point following
-
-        Busy waits until the boat moved to the point (bad, should be fixed with asyncio patterns)
-
-        Returns true if exit condition is met (exit_func)
-        Sends new waypoint if desired by the goal_update_func
-        - goal_update_func() -> should_update, (new_goal.x, new_goal.y)
-        '''
-        self.get_logger().info(f"Moving to point {point}")
-        self.moved_to_point = False
-        self.waypoint_rejected = False
-        self.waypoint_aborted = False
-        self.follow_path_client.wait_for_server()
-        goal_msg = FollowPath.Goal()
-        goal_msg.planner = self.get_parameter("planner").value
-        goal_msg.x = point[0]
-        goal_msg.y = point[1]
-        goal_msg.xy_threshold = self.get_parameter("xy_threshold").value
-        goal_msg.theta_threshold = self.get_parameter("theta_threshold").value
-        goal_msg.goal_tol = self.get_parameter("goal_tol").value
-        goal_msg.obstacle_tol = self.get_parameter("obstacle_tol").value
-        goal_msg.choose_every = self.get_parameter("choose_every").value
-        goal_msg.is_stationary = is_stationary
-
-        self._send_goal(goal_msg)
-        if busy_wait:
-            while not self.moved_to_point:
-                if (exit_func is not None) and exit_func():
-                    return True
-                if (goal_update_func is not None):
-                    update_goal, new_goal = goal_update_func()
-                    if update_goal:
-                        goal_msg.x = new_goal[0]
-                        goal_msg.y = new_goal[1]
-                        self.get_logger().info('ADAPTING GOAL POINT')
-                        self._send_goal(goal_msg)
-                if self.waypoint_rejected or self.waypoint_aborted:  # Retry functionality
-                    self.get_logger().info('RESENDING GOAL')
-                    self._send_goal(goal_msg)
-                    self.waypoint_rejected = False
-                    self.waypoint_aborted = False
-                time.sleep(self.timer_period)
-        return False
-
-    def follow_path_response_cb(self, future):
-        '''
-        Responds to follow path action server goal response.
-        '''
-        goal_handle = future.result()
-        self.active_future_request -= 1
-        if self.active_future_request > 0:
-            self.get_logger().info('PREVIOUS FUTURE, IGNORING')
-            return
-        
-        if not goal_handle.accepted:
-            self.get_logger().info('Waypoint rejected')
-            self.waypoint_rejected = True
-            return
-
-        self.get_logger().info("Waypoint accepted")
-        self._get_result_future = goal_handle.get_result_async()
-        self.active_waypoint_request += 1
-        self._get_result_future.add_done_callback(self.get_point_result_cb)
-
-    def get_point_result_cb(self, future):
-        '''
-        Flags the path following as complete for move_to_point
-        '''
-        self.active_waypoint_request -= 1
-        if self.active_waypoint_request > 0:
-            self.get_logger().info('PREVIOUS WAYPOINT, IGNORING')
-            return
-        
-        # Marks path following as finished/ moved to path following point
-        # if path following is interrupted, does not affect moved to point
-        result = future.result().result
-        status = future.result().status
-        if status == GoalStatus.STATUS_ABORTED:
-            self.get_logger().info(f'WAYPOINT ABORTED')
-            self.waypoint_aborted = True
-        elif result.is_finished:
-            self.get_logger().info(f'MOVED TO WAYPOINT')
-            self.moved_to_point = True
 
     def norm_squared(self, vec, ref=(0, 0)):
         return (vec[0] - ref[0])**2 + (vec[1]-ref[1])**2
