+++ all_seaing_common/all_seaing_common/task_server_base.py
@@ -1,10 +1,11 @@
 import rclpy
-from rclpy.action import ActionServer, CancelResponse, GoalResponse 
+from rclpy.action import ActionServer, CancelResponse, GoalResponse, ActionClient
 from rclpy.executors import MultiThreadedExecutor
 
 
-from all_seaing_interfaces.action import Task
+from all_seaing_interfaces.action import FollowPath, Task, Waypoint
 from all_seaing_common.action_server_base import ActionServerBase
+from action_msgs.msg import GoalStatus
 
 import os
 import yaml
@@ -25,11 +26,32 @@ class TaskServerBase(ActionServerBase):
             goal_callback=self.goal_callback
         )
 
+        self.follow_path_client = ActionClient(self, FollowPath, "follow_path")
+        self.waypoint_client = ActionClient(self, Waypoint, "waypoint")
+
         self.timer_period = (
             self.declare_parameter("timer_period", 1.0 / 30.0)
             .get_parameter_value()
             .double_value
         )
+
+        self.declare_parameter("xy_threshold", 1.0)
+        self.declare_parameter("theta_threshold", 180.0)
+        self.declare_parameter("wpt_theta_threshold", 10.0)
+        self.declare_parameter("goal_tol", 1.0)
+        self.declare_parameter("obstacle_tol", 50)
+        self.declare_parameter("choose_every", 10)
+        self.declare_parameter("planner", "astar")
+
+        self.declare_parameter("bypass_planner", False)
+
+        self.bypass_planner = self.get_parameter("bypass_planner").get_parameter_value().bool_value
+        
+        self.moved_to_point = False
+        self.waypoint_rejected = False
+        self.waypoint_aborted = False
+        self.active_future_request = 0 # active future requests
+        self.active_waypoint_request = 0 # keeps track of the number of active waypoint requests
 
     # Mark result of task as successful, exit control loop
     # IN THEORY, supports directly marking .result as true, and will mark _succeed as true by default
@@ -122,3 +144,158 @@ class TaskServerBase(ActionServerBase):
         self.end_process(f"Task Server [{self.server_name}] task completed with result {self._succeed}")
         goal_handle.succeed()
         return Task.Result(success=self._succeed)
+    
+    def _send_goal(self, goal_msg):
+        self.follow_path_client.wait_for_server()
+        self.active_future_request += 1
+        self.send_goal_future = self.follow_path_client.send_goal_async(
+            goal_msg
+        )
+        self.send_goal_future.add_done_callback(self.follow_path_response_cb)
+
+    def _send_wpt_goal(self, goal_msg):
+        self.waypoint_client.wait_for_server()
+        self.active_future_request += 1
+        self.send_goal_future = self.waypoint_client.send_goal_async(
+            goal_msg
+        )
+        self.send_goal_future.add_done_callback(self.follow_path_response_cb)
+
+    def move_to_point(self, point, is_stationary=False, busy_wait=False, exit_func=None, goal_update_func=None, cancel_on_exit=False):
+        '''
+        Moves the boat to the specified position using the follow path action server.
+        If busy_wait=true, then the returned truth value indicates success of point following
+
+        Busy waits until the boat moved to the point (bad, should be fixed with asyncio patterns)
+
+        Returns true if exit condition is met (exit_func)
+        Sends new waypoint if desired by the goal_update_func
+        - goal_update_func() -> should_update, (new_goal.x, new_goal.y)
+        '''
+        self.get_logger().info(f"Moving to point {point}")
+        self.moved_to_point = False
+        self.waypoint_rejected = False
+        self.waypoint_aborted = False
+        self.follow_path_client.wait_for_server()
+        goal_msg = FollowPath.Goal()
+        goal_msg.planner = self.get_parameter("planner").value
+        goal_msg.x = point[0]
+        goal_msg.y = point[1]
+        goal_msg.xy_threshold = self.get_parameter("xy_threshold").value
+        goal_msg.theta_threshold = self.get_parameter("theta_threshold").value
+        goal_msg.goal_tol = self.get_parameter("goal_tol").value
+        goal_msg.obstacle_tol = self.get_parameter("obstacle_tol").value
+        goal_msg.choose_every = self.get_parameter("choose_every").value
+        goal_msg.is_stationary = is_stationary
+
+        self._send_goal(goal_msg)
+        if busy_wait:
+            while not self.moved_to_point:
+                if (exit_func is not None) and exit_func():
+                    if cancel_on_exit:
+                        self.cancel_navigation()
+                    return True
+                if (goal_update_func is not None):
+                    update_goal, new_goal = goal_update_func()
+                    if update_goal:
+                        goal_msg.x = new_goal[0]
+                        goal_msg.y = new_goal[1]
+                        self.get_logger().info('ADAPTING GOAL POINT')
+                        self._send_goal(goal_msg)
+                if self.waypoint_rejected or self.waypoint_aborted:  # Retry functionality
+                    self.get_logger().info('RESENDING GOAL')
+                    self._send_goal(goal_msg)
+                    self.waypoint_rejected = False
+                    self.waypoint_aborted = False
+                time.sleep(self.timer_period)
+        return False
+
+    def follow_path_response_cb(self, future):
+        '''
+        Responds to follow path action server goal response.
+        '''
+        self.wpt_goal_handle = future.result()
+        self.active_future_request -= 1
+        if self.active_future_request > 0:
+            self.get_logger().info('PREVIOUS FUTURE, IGNORING')
+            return
+        
+        if not self.wpt_goal_handle.accepted:
+            self.get_logger().info('Waypoint rejected')
+            self.waypoint_rejected = True
+            return
+
+        self.get_logger().info("Waypoint accepted")
+        self._get_result_future = self.wpt_goal_handle.get_result_async()
+        self.active_waypoint_request += 1
+        self._get_result_future.add_done_callback(self.get_point_result_cb)
+
+    def cancel_navigation(self):
+        self.get_logger().info('TRYING TO CANCEL NAVIGATION')
+        if self._get_result_future is None or self.wpt_goal_handle is None: # TODO how to handle waypoint sent (& goal handle received) after us trying to cancel it? maybe ROS delay + retry for 3 times or smth
+            self.get_logger().info('FUTURE IS NONE')
+            return
+        if ((self._get_result_future.result() is None) or 
+            (self._get_result_future.result().status not in [GoalStatus.STATUS_ABORTED, GoalStatus.STATUS_SUCCEEDED, GoalStatus.STATUS_CANCELED, GoalStatus.STATUS_CANCELING])):
+            self.wpt_goal_handle.cancel_goal_async()
+            self.get_logger().info('CANCELLED NAVIGATION')
+
+    def get_point_result_cb(self, future):
+        '''
+        Flags the path following as complete for move_to_point
+        '''
+        self.active_waypoint_request -= 1
+        if self.active_waypoint_request > 0:
+            self.get_logger().info('PREVIOUS WAYPOINT, IGNORING')
+            return
+        
+        # Marks path following as finished/ moved to path following point
+        # if path following is interrupted, does not affect moved to point
+        result = future.result().result
+        status = future.result().status
+        if status == GoalStatus.STATUS_ABORTED:
+            self.get_logger().info(f'WAYPOINT ABORTED')
+            self.waypoint_aborted = True
+        elif result.is_finished:
+            self.get_logger().info(f'MOVED TO WAYPOINT')
+            self.moved_to_point = True
+
+    def move_to_waypoint(self, point, is_stationary=False, busy_wait=False, exit_func=None, goal_update_func=None, cancel_on_exit=False):
+        self.get_logger().info(f"Moving to waypoint {point}")
+        self.moved_to_point = False
+        goal_msg = Waypoint.Goal()
+        goal_msg.xy_threshold = self.get_parameter("xy_threshold").value
+        goal_msg.theta_threshold = self.get_parameter("wpt_theta_threshold").value
+        goal_msg.x = point[0]
+        goal_msg.y = point[1]
+        if len(point) >= 3:
+            goal_msg.theta = point[2]
+            goal_msg.ignore_theta = False
+        else:
+            goal_msg.ignore_theta = True
+        goal_msg.is_stationary = is_stationary
+        self._send_wpt_goal(goal_msg)
+        if busy_wait:
+            while not self.moved_to_point:
+                if (exit_func is not None) and exit_func():
+                    if cancel_on_exit:
+                        self.cancel_navigation()
+                    return True
+                if (goal_update_func is not None):
+                    update_goal, new_goal = goal_update_func()
+                    if update_goal:
+                        goal_msg.x = new_goal[0]
+                        goal_msg.y = new_goal[1]
+
+                        self.get_logger().info('ADAPTING WAYPOINT')
+                        self._send_wpt_goal(goal_msg)
+                time.sleep(self.timer_period)
+        return False
+
+    def send_waypoint_to_server(self, waypoint, is_stationary=False):
+        self.sent_waypoint = waypoint
+
+        if not self.bypass_planner:
+            self.move_to_point(waypoint, is_stationary=is_stationary)
+        else:
+            self.move_to_waypoint(waypoint, is_stationary=is_stationary)
